requirements.txt
langchain>=0.3
langchain-openai>=0.3
langchain-core>=0.3
lark-oapi>=1.4
httpx>=0.28
python-dotenv>=1.0
